- Make handstands so that SELECTs like the one in
    create table foo (i INTEGER);
    insert into foo values (NULL);
    insert into foo values (25);
    select * from foo UNION select * from foo order by i asc;

produce the expected type, i. e. an int here. The problem was that the first
row returns a NULL, and SQLite returns SQLITE_TEXT as the type for it, and we
only looked at the first row for detecting the type. The algorithm is now
changed that when the value in the first row is NULL, then the type code for
this column is set to None (this is backwards-incompatible, but working around
means rewriting half of _sqlite.c, so that's ruled out). Also, for this column,
the type is determined in each call of process_record and the appropriate
Python value is created.

- Actually detect errors after sqlite3_prepare, and abort if any occured.


--HG--
branch : 1.1
extra : convert_revision : svn%3Af581f19b-c9e6-0310-bd56-e372eea7194d/pysqlite/branches/1.1%4043